Configure access to lighthouse through the proxy
In order to work around how nginx handles encoded urls (when it needs to
rewrite it decodes them and passes them on unencoded) we need to access
lighthouse via the backend. This was recommended in passing by the spotify
team anyway. It seems express doesn't behave in the same way.

To make use of this set appConfig.lighthouse.baseUrl to:
 https://<public_url>/api/proxy/lighthouse