Update csp policy for helm charts & fix ingress for lighthouse
The lighthouse report is loaded as html which has embedded scripts
so add 'unsafe-inline' to the CSP in order to view audits in the
helm chart deployments.

Remove the lighthouse path based ingress. Nginx path rewriting
decodes the website url and passes it on unencoded which doesn't work.
The solution is to forward requests using the backend proxy as the
backend ingress doesn't rewrite urls.